Мелкофикс педальной системы

+++ code/modules/assembly/infrared.dm
@@ -200,11 +200,11 @@
 
 /obj/effect/beam/i_beam/process()
 	//world << "i_beam \ref[src] : process"
-
-	if((loc.density || !(master)))
+	if(loc)
+		if((loc.density || !(master)))
 	//	world << "beam hit loc [loc] or no master [master], deleting"
-		qdel(src)
-		return
+			qdel(src)
+			return
 	//world << "proccess: [src.left] left"
 
 	if(left > 0)
